backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0
python-multipart>=0.0.5
aiofiles>=23.2.1

# Database
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
alembic>=1.7.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT==2.6.0
pydantic[email]>=1.8.0

# HTTP and utilities
requests>=2.26.0
httpx>=0.24.0
aiohttp>=3.12.0
cors>=1.0.1

# Document processing (simplified)
python-docx>=0.8.11

# Image processing
Pillow>=10.0.0
filetype>=1.2.0

# AI integration
openai>=1.0.0
pydantic>=2.0.0

# Note: For best results, install these system tools:
# - Pandoc: https://pandoc.org/installing.html (90% accuracy)
# - LibreOffice: https://www.libreoffice.org/download/ (90% accuracy)
# Without these, the system falls back to basic python-docx (50% accuracy)